use a proper highlighted_snippet_for_string implementation
I previously switched us to the newlines mode, but forgot about this function
requiring the nonewlines mode syntaxes. This corrects that by reimplementing it
for the newlines mode. This was the cause of e.g. never ending' Go comments
when highlighting Go files, and probably a slew of other bugs.