Add rust toolchain file
Why? 

This app requires rocket which works on nightly. Nightly build are released everyday so it can break anyday. But when developer do it they usually test on one nightly and other contributor can pull this project and bam it don't work.

If we specify the nightly version in this file cargo etc will download the toolchain specified here and use that nightly instead of current nightly version.