fix Dockerfile / musl compilation
After upgrading to Rust stable this Dockerfile would produce an invalid image
which segfaulted. I am not sure of the exact cause, but new rust `-alpine`
variants now exist (they did not before) which serve to perform this cross-compilation
in an easier way. Using that, a valid binary is produced.