Add code for highlighting with CSS
This adds some code to emit CSS-classed spans rather than in-line
styles. This allows us to leave the theming to the browser.

Additionally, this emits the line-numbered table as we use it in the
frontend so we don't need to parse and modify the returned HTML. This
significantly simplifies the logic of the client.