Revert "Trim newlines from highlighted line"
This reverts commit 7766dd1539f0e058961ad69ad3cc32f49ce848b4.  It turns
out, wrapping the content in a <div> (as was done a few commits earlier)
is sufficient to fix the issue we were seeing with codecov badges. By
skipping the newline trimming, there is less complexity, and it adds the
bonus that the textContent property of the node is the exact file
contents. Note that this does not affect copy/paste.